requirements.txt
moviepy==1.0.3
numpy==1.21.5
opencv_python==4.5.4.60
Pillow==8.4.0
pygame==2.1.1
tqdm==4.62.3
